fix a bug where two update alerts appear. Change the section from chat rooms to be only for days.
git-svn-id: svn://mobilep.ro/status@220 99022451-f8c5-4652-be14-efb92feb6bba

+++ trunk/Status/Categories/NSDate+Additions.m
@@ -55,7 +55,7 @@
 //    NSInteger months = [components month];
 //    NSInteger days = [components day];
     
-    return [NSString stringWithFormat:@"%li", years];
+    return [NSString stringWithFormat:@"%li", (long)years];
 }
 
 @end
